Start on update to Sesame-4

+++ core/src/main/java/org/apache/any23/extractor/rdf/RDFParserFactory.java
@@ -265,7 +265,7 @@ public class RDFParserFactory {
             extractionResult = er;
         }
 
-        public void warning(String msg, int lineNo, int colNo) {
+        public void warning(String msg, long lineNo, long colNo) {
             try {
                 extractionResult.notifyIssue(IssueReport.IssueLevel.Warning, msg, lineNo, colNo);
             } catch (Exception e) {
@@ -273,7 +273,7 @@ public class RDFParserFactory {
             }
         }
 
-        public void error(String msg, int lineNo, int colNo) {
+        public void error(String msg, long lineNo, long colNo) {
             try {
                 extractionResult.notifyIssue(IssueReport.IssueLevel.Error, msg, lineNo, colNo);
             } catch (Exception e) {
@@ -281,7 +281,7 @@ public class RDFParserFactory {
             }
         }
 
-        public void fatalError(String msg, int lineNo, int colNo) {
+        public void fatalError(String msg, long lineNo, long colNo) {
             try {
                 extractionResult.notifyIssue(IssueReport.IssueLevel.Fatal, msg, lineNo, colNo);
             } catch (Exception e) {
